+++ YubiKit/YubiKit/Connections/QRReaderSession/YKFQRCodeScanViewController.m
@@ -168,8 +168,6 @@
     
     AudioServicesPlaySystemSound(kSystemSoundID_Vibrate);
     [self.delegate qrCodeScanViewController:self didScanPayload:payload];
-    
-    YKFLogInfo(@"QR code scanned with value: %@", payload);
 }
 
 #pragma mark - UI Setup