+++ YubiKit/YubiKit/Connections/Shared/Sessions/FIDO2/YKFFIDO2Session.m
@@ -76,8 +76,6 @@ typedef void (^YKFFIDO2SessionClientPinSharedSecretCompletionBlock)
 // Keeps the state of the application selection to avoid reselecting the application.
 @property BOOL applicationSelected;
 
-@property (nonatomic, readwrite) YKFSmartCardInterface *smartCardInterface;
-
 @end
 
 @implementation YKFFIDO2Session
@@ -255,8 +253,10 @@ typedef void (^YKFFIDO2SessionClientPinSharedSecretCompletionBlock)
 - (void)setPin:(nonnull NSString *)pin completion:(nonnull YKFFIDO2SessionGenericCompletionBlock)completion {
     YKFParameterAssertReturn(pin);
     YKFParameterAssertReturn(completion);
-
-    if (pin.length < 4 || pin.length > 255) {
+    
+    NSData *pinData = [[pin dataUsingEncoding:NSUTF8StringEncoding] ykf_fido2PaddedPinData];
+    
+    if (pin.length < 4 || pinData.length > 65) {
         completion([YKFFIDO2Error errorWithCode:YKFFIDO2ErrorCodePIN_POLICY_VIOLATION]);
         return;
     }
@@ -277,7 +277,6 @@ typedef void (^YKFFIDO2SessionClientPinSharedSecretCompletionBlock)
         setPinRequest.subCommand = YKFFIDO2ClientPinRequestSubCommandSetPIN;
         setPinRequest.keyAgreement = cosePlatformPublicKey;
         
-        NSData *pinData = [[pin dataUsingEncoding:NSUTF8StringEncoding] ykf_fido2PaddedPinData];
         
         setPinRequest.pinEnc = [pinData ykf_aes256EncryptedDataWithKey:sharedSecret];
         setPinRequest.pinAuth = [[setPinRequest.pinEnc ykf_fido2HMACWithKey:sharedSecret] subdataWithRange:NSMakeRange(0, 16)];