+++ YubiKit/YubiKit/Connections/Shared/Sessions/PIV/YKFPIVSession.h
@@ -16,6 +16,7 @@
 #define YKFPIVSession_h
 
 #import "YKFVersion.h"
+#import "YKFSession+Private.h"
 #import "YKFPIVKeyType.h"
 
 /// Touch policy for PIV application.
@@ -51,7 +52,11 @@ typedef NS_ENUM(NSUInteger, YKFPIVFErrorCode) {
     YKFPIVFErrorCodeInvalidCipherTextLength = 1,
     YKFPIVFErrorCodeUnsupportedOperation = 2,
     YKFPIVFErrorCodeDataParseError = 3,
-    YKFPIVFErrorCodeUnknownKeyType = 4
+    YKFPIVFErrorCodeUnknownKeyType = 4,
+    YKFPIVFErrorCodeInvalidPin = 5,
+    YKFPIVFErrorCodePinLocked = 6,
+    YKFPIVFErrorCodeInvalidResponse = 7,
+    YKFPIVFErrorCodeAuthenticationFailed = 8
 };
 
 @class YKFPIVSessionFeatures, YKFPIVManagementKeyType, YKFPIVManagementKeyMetadata;
@@ -145,7 +150,7 @@ typedef void (^YKFPIVSessionManagementKeyMetadataCompletionBlock)
 /// @abstract Provides the interface for executing PIV requests with the key.
 /// @discussion The PIV session is mantained by the YKFConnection which controls its lifecycle. The application
 ///             must not create one.
-@interface YKFPIVSession: NSObject <YKFVersionProtocol>
+@interface YKFPIVSession: YKFSession <YKFVersionProtocol>
 
 /// @abstract This property provides the version of the currently connected YubiKey.
 @property (nonatomic, readonly) YKFVersion * _Nonnull version;