+++ YubiKit/YubiKit/YubiKitManager.h
@@ -77,6 +77,19 @@
   */
 - (void)didDisconnectAccessory:(YKFAccessoryConnection *_Nonnull)connection error:(NSError *_Nullable)error;
 
+/*!
+ @method didFailConnectingNFC
+ 
+ @abstract
+    The YubiKey SDK did receive a NFC connection error. This is typically that the user pressed cancel (error code 200) in the
+    NFC modal or it timed out (error code 201) waiting for a NFC YubiKey.
+
+ @method error
+    The NSError passed by the SDK.
+  */
+@optional
+- (void)didFailConnectingNFC:(NSError *_Nonnull)error;
+
 @end
 
 
@@ -93,7 +106,9 @@
  
  @abstract
     The delegate must conform to the YKFManagerDelegate protocol. Setting this delegate will allow you
-    to get notifications when a connection to the YubiKey is established or broken.
+    to get notifications when a connection to the YubiKey is established or broken. If a connection is
+    already established when the delegate is assigned the didConnect delegate methods will be called
+    immediately.
  */
 @property(nonatomic, weak) id<YKFManagerDelegate> _Nullable delegate;
 
@@ -116,6 +131,30 @@
     will also dismiss the NFC system modal presented by iOS during NFC operations.
  */
 - (void)stopNFCConnection API_AVAILABLE(ios(13.0));
+
+/*!
+ @method stopNFCConnectionWithMessage:
+ 
+ @abstract
+    Stop the NFC connection and display a message.
+ 
+ @discussion
+    Use this method to close the NFC connection and display a message to the user.
+  */
+- (void)stopNFCConnectionWithMessage:(NSString *_Nonnull)message API_AVAILABLE(ios(13.0)) NS_SWIFT_NAME(stopNFCConnection(withMessage:));
+
+/*!
+ @method stopNFCConnectionWithErrorMessage:
+ 
+ @abstract
+    Stop the NFC connection and display a error message.
+ 
+ @discussion
+    Use this method to close the NFC connection and display a message to the user when an error condition
+    occurs after the app successfully executed a command. This type of error can occur if, for example,
+    you send a password to unlock the YubiKey and it is not the matching password.
+  */
+- (void)stopNFCConnectionWithErrorMessage:(NSString *_Nonnull)errorMessage API_AVAILABLE(ios(13.0)) NS_SWIFT_NAME(stopNFCConnection(withErrorMessage:));
 
 /*!
  @method startAccessoryConnection